Refactor machine state deleter
