+++ packages/arb-avm-cpp/data_storage/src/checkpoint/machinestatedeleter.cpp
@@ -20,6 +20,20 @@
 #include <data_storage/storageresult.hpp>
 #include <data_storage/transaction.hpp>
 
+class MachineStateDeleter {
+   private:
+    std::unique_ptr<Transaction> transaction;
+    DeleteResults deleteTuple(const std::vector<unsigned char>& hash_key,
+                              GetResults results);
+
+   public:
+    MachineStateDeleter(std::unique_ptr<Transaction> transaction_);
+    DeleteResults deleteTuple(const std::vector<unsigned char>& hash_key);
+    DeleteResults deleteValue(const std::vector<unsigned char>& hash_key);
+    rocksdb::Status commitTransaction();
+    rocksdb::Status rollBackTransaction();
+};
+
 MachineStateDeleter::MachineStateDeleter(
     std::unique_ptr<Transaction> transaction_) {
     transaction = std::move(transaction_);
@@ -28,22 +42,22 @@ MachineStateDeleter::MachineStateDeleter(
 DeleteResults MachineStateDeleter::deleteTuple(
     const std::vector<unsigned char>& hash_key,
     GetResults results) {
-    if (results.status.ok()) {
-        if (results.reference_count == 1) {
-            auto value_vectors =
-                checkpoint::utils::parseTuple(results.stored_value);
-
-            for (auto& vector : value_vectors) {
-                if (static_cast<ValueTypes>(vector[0]) == TUPLE) {
-                    vector.erase(vector.begin());
-                    auto delete_status = deleteTuple(vector);
-                }
-            }
-        }
-        return transaction->deleteData(hash_key);
-    } else {
+    if (!results.status.ok()) {
         return DeleteResults{0, results.status};
     }
+
+    if (results.reference_count == 1) {
+        auto value_vectors =
+            checkpoint::utils::parseTuple(results.stored_value);
+
+        for (auto& vector : value_vectors) {
+            if (static_cast<ValueTypes>(vector[0]) == TUPLE) {
+                vector.erase(vector.begin());
+                auto delete_status = deleteTuple(vector);
+            }
+        }
+    }
+    return transaction->deleteData(hash_key);
 }
 
 DeleteResults MachineStateDeleter::deleteTuple(
@@ -56,16 +70,16 @@ DeleteResults MachineStateDeleter::deleteValue(
     const std::vector<unsigned char>& hash_key) {
     auto results = transaction->getData(hash_key);
 
-    if (results.status.ok()) {
-        auto type = static_cast<ValueTypes>(results.stored_value[0]);
-
-        if (type == TUPLE) {
-            return deleteTuple(hash_key, results);
-        } else {
-            return transaction->deleteData(hash_key);
-        }
-    } else {
+    if (!results.status.ok()) {
         return DeleteResults{0, results.status};
+    }
+
+    auto type = static_cast<ValueTypes>(results.stored_value[0]);
+
+    if (type == TUPLE) {
+        return deleteTuple(hash_key, results);
+    } else {
+        return transaction->deleteData(hash_key);
     }
 }
 
@@ -83,34 +97,38 @@ DeleteResults deleteCheckpoint(
     auto results = checkpoint_storage.getValue(checkpoint_name);
     auto deleter = MachineStateDeleter(checkpoint_storage.makeTransaction());
 
-    if (results.status.ok()) {
-        auto delete_results = deleter.deleteValue(checkpoint_name);
-
-        if (delete_results.reference_count < 1) {
-            auto parsed_state =
-                checkpoint::utils::extractStateKeys(results.stored_value);
-
-            auto delete_register_res =
-                deleter.deleteValue(parsed_state.register_val_key);
-            auto delete_cp_key = deleter.deleteValue(parsed_state.pc_key);
-            auto delete_err_pc = deleter.deleteValue(parsed_state.err_pc_key);
-            auto delete_datastack_res =
-                deleter.deleteTuple(parsed_state.datastack_key);
-            auto delete_auxstack_res =
-                deleter.deleteTuple(parsed_state.auxstack_key);
-
-            if (!(delete_register_res.status.ok() &&
-                  delete_cp_key.status.ok() &&
-                  delete_datastack_res.status.ok() &&
-                  delete_auxstack_res.status.ok() &&
-                  delete_err_pc.status.ok())) {
-                std::cout << "error deleting checkpoint" << std::endl;
-            }
-        }
-        auto status = deleter.commitTransaction();
-        delete_results.status = status;
-        return delete_results;
-    } else {
+    if (!results.status.ok()) {
         return DeleteResults{0, results.status};
     }
+
+    auto delete_results = deleter.deleteValue(checkpoint_name);
+
+    if (delete_results.reference_count < 1) {
+        auto parsed_state =
+            checkpoint::utils::extractStateKeys(results.stored_value);
+
+        auto delete_register_res =
+            deleter.deleteValue(parsed_state.register_val_key);
+        auto delete_cp_key = deleter.deleteValue(parsed_state.pc_key);
+        auto delete_err_pc = deleter.deleteValue(parsed_state.err_pc_key);
+        auto delete_datastack_res =
+            deleter.deleteTuple(parsed_state.datastack_key);
+        auto delete_auxstack_res =
+            deleter.deleteTuple(parsed_state.auxstack_key);
+
+        if (!(delete_register_res.status.ok() && delete_cp_key.status.ok() &&
+              delete_datastack_res.status.ok() &&
+              delete_auxstack_res.status.ok() && delete_err_pc.status.ok())) {
+            std::cout << "error deleting checkpoint" << std::endl;
+        }
+    }
+    auto status = deleter.commitTransaction();
+    delete_results.status = status;
+    return delete_results;
+}
+
+DeleteResults deleteValue(CheckpointStorage& checkpoint_storage,
+                          const std::vector<unsigned char>& hash_key) {
+    auto deleter = MachineStateDeleter(checkpoint_storage.makeTransaction());
+    return deleter.deleteValue(hash_key);
 }
